chore: Moves Chart to the components folder
