+++ superset-frontend/src/dashboard/actions/hydrate.js
@@ -24,7 +24,7 @@ import {
   getChartMetadataRegistry,
 } from '@superset-ui/core';
 
-import { chart } from 'src/chart/chartReducer';
+import { chart } from 'src/components/Chart/chartReducer';
 import { initSliceEntities } from 'src/dashboard/reducers/sliceEntities';
 import { getInitialState as getInitialNativeFilterState } from 'src/dashboard/reducers/nativeFilters';
 import { applyDefaultFormData } from 'src/explore/store';
