+++ superset-frontend/src/dashboard/components/FiltersBadge/FiltersBadge.test.tsx
@@ -26,7 +26,7 @@ import { styledMount as mount } from 'spec/helpers/theming';
 import {
   CHART_RENDERING_SUCCEEDED,
   CHART_UPDATE_SUCCEEDED,
-} from 'src/chart/chartAction';
+} from 'src/components/Chart/chartAction';
 import { buildActiveFilters } from 'src/dashboard/util/activeDashboardFilters';
 import { FiltersBadge } from 'src/dashboard/components/FiltersBadge';
 import {
