+++ superset-frontend/src/dashboard/types.ts
@@ -25,7 +25,7 @@ import {
   NativeFiltersState,
 } from '@superset-ui/core';
 import { DatasourceMeta } from '@superset-ui/chart-controls';
-import { chart } from 'src/chart/chartReducer';
+import { chart } from 'src/components/Chart/chartReducer';
 import componentTypes from 'src/dashboard/util/componentTypes';
 
 import { User } from 'src/types/bootstrapTypes';
