+++ superset-frontend/src/explore/components/DataTablesPane/index.tsx
@@ -29,7 +29,7 @@ import Tabs from 'src/components/Tabs';
 import Loading from 'src/components/Loading';
 import { EmptyStateMedium } from 'src/components/EmptyState';
 import TableView, { EmptyWrapperType } from 'src/components/TableView';
-import { getChartDataRequest } from 'src/chart/chartAction';
+import { getChartDataRequest } from 'src/components/Chart/chartAction';
 import { getClientErrorObject } from 'src/utils/getClientErrorObject';
 import {
   getItem,
