+++ superset-frontend/src/explore/components/ExploreChartPanel.jsx
@@ -22,7 +22,7 @@ import Split from 'react-split';
 import { styled, SupersetClient, useTheme } from '@superset-ui/core';
 import { useResizeDetector } from 'react-resize-detector';
 import { chartPropShape } from 'src/dashboard/util/propShapes';
-import ChartContainer from 'src/chart/ChartContainer';
+import ChartContainer from 'src/components/Chart/ChartContainer';
 import {
   getItem,
   setItem,
