+++ superset-frontend/src/explore/reducers/exploreReducer.js
@@ -18,7 +18,7 @@
  */
 /* eslint camelcase: 0 */
 import { ensureIsArray } from '@superset-ui/core';
-import { DYNAMIC_PLUGIN_CONTROLS_READY } from 'src/chart/chartAction';
+import { DYNAMIC_PLUGIN_CONTROLS_READY } from 'src/components/Chart/chartAction';
 import { DEFAULT_TIME_RANGE } from 'src/explore/constants';
 import { getControlsState } from 'src/explore/store';
 import {
